requirements.txt
django==4.1.2
django-extensions==3.2.1
djangorestframework==3.14.0
flake8==5.0.4
flake8-builtins==2.0.0
flake8-import-order==0.18.1
flake8-polyfill==1.0.2
psycopg2-binary==2.9.4
django-ckeditor==6.5.1
